further page restructuring to allow for multiple service types and instances
