+++ src/App.js
@@ -1,23 +1,21 @@
+import React from 'react';
 import './App.css';
 import '@fontsource/roboto';
-import React, { useEffect } from 'react';
 import {
   BrowserRouter as Router,
   Switch,
   Route
 } from "react-router-dom";
 import { ThemeProvider } from '@material-ui/styles';
-import Landing from './lib/components/pages/Landing';
-import DrsMain from './lib/components/drs/DrsMain';
-import Main from './lib/components/pages/Main';
+import {
+    Landing,
+    Main
+} from './lib/components/pages';
 import theme from './lib/styles/theme/theme';
 
 const App = () => {
-  
     return (
-        //<ThemeProvider theme={theme}>
         <ThemeProvider theme={theme}>
-        
             <Router>
                 <Switch>
                     <Route exact path='/' component={Landing} />
