+++ src/lib/components/ga4gh/drs/drsobject/formComponents/DrsObjectRelatives.js
@@ -12,7 +12,7 @@ import {
 import {
     Link
 } from 'react-router-dom';
-import { scrollToTop } from '../../../../functions/common';
+import { scrollToTop } from '../../../../../functions/common';
 import SpaceDivider from '../../../../common/SpaceDivider';
 import RemoveItemButton from '../../../../common/RemoveItemButton';
 import VerifyIdButton from './VerifyIdButton';
