+++ src/lib/components/pages/Main.js
@@ -7,7 +7,7 @@ import About from './About';
 import Home from './Home';
 import StarterKitAppBar from '../common/StarterKitAppBar';
 import StarterKitBottomNav from '../common/StarterKitBottomNav';
-import DrsObjectMain from '../apis/drs/drsobject/DrsObjectMain';
+import DrsObjectMain from '../ga4gh/drs/drsobject/DrsObjectMain';
 import { makeStyles } from '@material-ui/core/styles';
 import Services from './Services';
 import Service from './Service';
@@ -53,25 +53,32 @@ const Main = () => {
                         <Services trail={servicesTrail} url='/services' />
                     </Route>
 
-                    {/*
-                        for each service in the config, render its service
-                        summary, as well as the 'Main' component for its API
-                        type
-                    */}
-
+                    {/* for each service in the config: */}
                     {hardCodedServiceConfigs.map(service => {
                         let apiType = ga4ghApiTypes[service.serviceType];
                         return (
                             <div>
+                                {/* render the service summary */}
                                 <Route exact path={`/services/${service.id}`}>
                                     <Service
                                         service={service}
                                         trail={servicesTrail}
                                     />
                                 </Route>
-                                <Route exact path={`/services/${service.id}/${service.serviceType}/objects`}>
-                                    {renderApiTypeComponent(service)}
-                                </Route>
+                                {apiType.models.map(model => {
+                                    let trail = [...servicesTrail];
+                                    trail.push({to: `/services/${service.id}`, label: service.id})
+                                    trail.push({to: `/services/${service.id}/${service.serviceType}`, label: service.serviceType});
+                                    trail.push({to: `/services/${service.id}/${service.serviceType}/${model.path}`, label: model.path});
+                                    return (
+                                        <Route
+                                            exact
+                                            path={`/services/${service.id}/${service.serviceType}/${model.path}`}
+                                        >
+                                            {model.componentFunction(service, trail)};
+                                        </Route>
+                                    )
+                                })}
                             </div>
                         )
                     })}
