+++ src/lib/components/pages/index.js
@@ -1,0 +1,15 @@
+import About from './About';
+import Home from './Home';
+import Landing from './Landing';
+import Main from './Main';
+import Service from './Service';
+import Services from './Services';
+
+export {
+    About,
+    Home,
+    Landing,
+    Main,
+    Service,
+    Services
+}
